chore: create `proxycore.scheduler` module
