--- vane-proxy-core/src/main/java/org/oddlama/vane/proxycore/ScheduledTask.java
@@ -1,11 +1,0 @@
-package org.oddlama.vane.proxycore;
-
-public interface ScheduledTask {
-	int getId();
-
-	VaneProxyPlugin getOwner();
-
-	Runnable getTask();
-
-	void cancel();
-}
