+++ vane-proxy-core/src/main/java/org/oddlama/vane/proxycore/scheduler/ProxyScheduledTask.java
@@ -1,0 +1,13 @@
+package org.oddlama.vane.proxycore.scheduler;
+
+import org.oddlama.vane.proxycore.VaneProxyPlugin;
+
+public interface ProxyScheduledTask {
+	int getId();
+
+	VaneProxyPlugin getOwner();
+
+	Runnable getTask();
+
+	void cancel();
+}
