+++ vane-waterfall/src/main/java/org/oddlama/vane/waterfall/compat/BungeeCompatProxyScheduledTask.java
@@ -2,11 +2,12 @@ package org.oddlama.vane.waterfall.compat;
 
 import net.md_5.bungee.api.scheduler.ScheduledTask;
 import org.oddlama.vane.proxycore.VaneProxyPlugin;
+import org.oddlama.vane.proxycore.scheduler.ProxyScheduledTask;
 
-public class BungeeCompatScheduledTask implements org.oddlama.vane.proxycore.ScheduledTask {
+public class BungeeCompatProxyScheduledTask implements ProxyScheduledTask {
 	public ScheduledTask task;
 
-	public BungeeCompatScheduledTask(ScheduledTask task) {
+	public BungeeCompatProxyScheduledTask(ScheduledTask task) {
 		this.task = task;
 	}
